+++ packages/flutter_meedu/test/listeners/provider_listener_state_provider_with_filters_test.dart
@@ -22,6 +22,7 @@ void main() {
           home: ProviderListener<_CounterController>(
             provider: _counterProvider.select(
               (_) => _.counter >= minValueToBeNotified,
+              booleanCallback: true,
             ),
             onChange: (_, __) {
               numberOfCalls++;