+++ packages/flutter_meedu/test/listeners/provider_listener_test_with_filters_test.dart
@@ -21,6 +21,7 @@ void main() {
           home: ProviderListener<_CounterController>(
             provider: _counterProvider.select(
               (_) => _.counter >= minValueToBeNotified,
+              booleanCallback: true,
             ),
             onChange: (_, __) {
               numberOfCalls++;