+++ packages/meedu/lib/src/rx/extensions.dart
@@ -1,21 +1,13 @@
 import 'rx.dart';
 
-/// creates an int observable using for example 0.obs
-extension RxInt on int {
-  Rx<int> get obs => Rx<int>(this);
-}
-
-/// creates a bool observable using for example false.obs
-extension RxBool on bool {
-  Rx<bool> get obs => Rx<bool>(this);
-}
-
-/// creates a double observable using for example 0.0.obs
-extension RxDouble on double {
-  Rx<double> get obs => Rx<double>(this);
-}
-
-/// creates a String observable using for example "".obs
-extension RxString on String {
-  Rx<String> get obs => Rx<String>(this);
+/// creates an int observable using the prefix .obx
+///
+/// examples:
+/// final query =  ''.obs;  // is equals to final Rx<String> query = Rx('');
+/// final loading =  true.obs;  // is equals to final Rx<bool> query = Rx(true);
+/// final price =  0.0.obs;  // is equals to final Rx<double> query = Rx(0.0);
+/// final counter =  0.obs;  // is equals to final Rx<int> query = Rx(0);
+/// final counter =  MyClass().obs;  // is equals to final Rx<MyClass> query = Rx(MyClass);
+extension RxExtension<T> on T {
+  Rx<T> get obs => Rx<T>(this);
 }