+++ website/docusaurus.config.js
@@ -75,7 +75,11 @@ module.exports = {
           lastVersion: "current",
           versions: {
             current: {
-              label: "7.1.0",
+              label: "8.0.0",
+            },
+            "7.x.x": {
+              label: "7.2.0",
+              path: "7.x.x",
             },
             "6.x.x": {
               label: "6.2.3",